User story MAT-579 Server Side Validation (Admin)
In AdminServiceImpl.searchUsers(), we are not checking if the request is coming from a user who is a
Admistrator. So we check the session to make sure that the user is indeed an Administrator. If not then AdminServiceImpl.searchUsers() will throw a mat.shared.InCorrectUserRoleException.

In ManageUsersPresenter (which makes the call to AdminServiceImpl.searchUsers() ), we check the onFailure(Throwable caught) method and if "caught" is instanceof InCorrectUserRoleException, then we automatically logout the user and show the login screen.

git-svn-id: svn+ssh://subversion.telligen.com/var/www/html/repos/cmat/trunk/Development/mat@363 de2f59c9-771b-4d1d-97da-b950f46c8601

Former-commit-id: 6309def75bd8b8e5a3fb13b9bd9371a12cc36d49