MAT-832:scroll bar doesn't appear on the change Password screen
The style property for the body tag "overflow:hidden" causes the scrollbar to be suppressed.
Removed the overflow:hidden and the Forgot password screen shows up with a vertical scrollbar.
The style property was added for setting a favicon. 
Tested and made sure that the favicon appears after "overflow:hidden" was removed.

git-svn-id: svn+ssh://subversion.telligen.com/var/www/html/repos/cmat/trunk/Development/mat@702 de2f59c9-771b-4d1d-97da-b950f46c8601

Former-commit-id: 20dea8188fd5963fa2d851c8cbc3ec9d8a74b023