MAT-8644:As a user, I would like the human readable output to be able to include the applicable datatype/value set combinations in the data criteria section that are used in retrieves when those retrieves are used in a measure.
MAT-8644:As a user, I would like the human readable output to be able to include the applicable datatype/value set combinations in the data criteria section that are used in retrieves when those retrieves are used in a measure.

Make the codes and value-sets so that they are sorted by name.

git-svn-id: svn+ssh://subversion.telligen.com/var/www/html/repos/cmat/trunk/Development/mat@15219 de2f59c9-771b-4d1d-97da-b950f46c8601

Former-commit-id: 9cf103b20d573c3fd149175b867404e50fbe5e88