MAT-8644:As a user, I would like the human readable output to be able to include the applicable datatype/value set combinations in the data criteria section that are used in retrieves when those retrieves are used in a measure.
MAT-8643:As a user, I would like the human readable output to be able to include the applicable datatype/direct reference code combinations in the data criteria section that are used in retrieves when those retrieves are used in a measure.

For Human readable data criteria, only consider QDM tags which have a "datatype" attribute.

git-svn-id: svn+ssh://subversion.telligen.com/var/www/html/repos/cmat/trunk/Development/mat@15225 de2f59c9-771b-4d1d-97da-b950f46c8601

Former-commit-id: 594b5cdcd1cc564a5776d3d2e256e3bf8aaf1fa9