MAT-8768: CQL Workspace delete messages - Definition/Function/Parameter/Argument/Includes ========================================================================================= CQLLibraryService: ================== As part of code cleanup, modified the following method signatures.
- deleteInclude
- deleteDefinition
- deleteFunctions
- deleteParameter

Removed an argument in the above methods that keeps the currently selected object, which is not required in the method.

CQLLibraryServiceAsync:
=======================
As part of code cleanup, modified the following method signatures.

- deleteInclude
- deleteDefinition
- deleteFunctions
- deleteParameter

Removed an argument in the above methods that keeps the currently selected object, which is not required in the method.

CQLService:
===========
As part of code cleanup, modified the following method signatures.

- deleteInclude
- deleteDefinition
- deleteFunctions
- deleteParameter

Removed an argument in the above methods that keeps the currently selected object, which is not required in the method.

MeasureService:
===============
As part of code cleanup, modified the following method signatures.

- deleteInclude
- deleteDefinition
- deleteFunctions
- deleteParameter

Removed an argument in the above methods that keeps the currently selected object, which is not required in the method.

MeasureServiceAsync:
====================
As part of code cleanup, modified the following method signatures.

- deleteInclude
- deleteDefinition
- deleteFunctions
- deleteParameter

Removed an argument in the above methods that keeps the currently selected object, which is not required in the method.



git-svn-id: svn+ssh://subversion.telligen.com/var/www/html/repos/cmat/trunk/Development/mat@15870 de2f59c9-771b-4d1d-97da-b950f46c8601

Former-commit-id: 38b0ed5646c72fe5e10f369955687a628360647f