MAT-8768: CQL Workspace delete messages - Definition/Function/Parameter/Argument/Includes =========================================================================================
CQLStandaloneWorkSpacePresenter:
================================
Modified the references in the below methods to get the new message from MessageDelegate
- deleteInclude
- deleteDefinition
- deleteFunctions
- deleteParameter
- deleteFunctionArguments

As part of code cleanup, in deleteDefinition, removed unnecessary code for validation of special characters and capturing the unused current object to be deleted.

As part of code cleanup, in deleteFunction, removed unnecessary code for validation and capturing the unused current object to be deleted.

As part of code cleanup, in deleteParameter, removed unnecessary code for validation and capturing the unused current object to be deleted.

As part of code cleanup, in deleteInclude, removed unnecessary code for validation and capturing the unused current object to be deleted.


CQLWorkSpacePresenter:
======================
Modified the references in the below methods to get the new message from MessageDelegate
- deleteInclude
- deleteDefinition
- deleteFunctions
- deleteParameter
- deleteFunctionArguments

As part of code cleanup, in deleteDefinition, removed unnecessary code for validation of special characters and capturing the unused current object to be deleted.

As part of code cleanup, in deleteFunction, removed unnecessary code for validation and capturing the unused current object to be deleted.

As part of code cleanup, in deleteParameter, removed unnecessary code for validation and capturing the unused current object to be deleted.

As part of code cleanup, in deleteInclude, removed unnecessary code for validation and capturing the unused current object to be deleted.


DeleteConfirmationDialogBox
===========================

Created a new method that will display the popup message in multiple lines passed in the ArrayList.

public void show(ArrayList<String> messageList)



git-svn-id: svn+ssh://subversion.telligen.com/var/www/html/repos/cmat/trunk/Development/mat@15871 de2f59c9-771b-4d1d-97da-b950f46c8601

Former-commit-id: cfd623d5b9122bfda49d23f923b26b687197d18f