MAT-9088.
1. CQLAppliedValueSetView.java - Added method setSelectedValueIndex(..)
to set the index value of selected value in a listbox.
2. Re-factored method setReleaseAndProgramAndVersionFieldsOnEdit(..) in
CQLStandaloneWorkSpacePresenter.java to use the method from 1 above to
set the Program and Release selection values. Disabled the version list
box on edit of value set.
3. Re-factored method setReleaseAndProgramAndVersionFieldsOnEdit(..) in
CQLWorkSpacePresenter.java to use the method from 1 above to set the
Program and Release selection values.


Former-commit-id: 0d28e918ac22c7d7bea67ddee32a6bcdb5218c13