MAT-10045: Changes to handle 401 unauthorized error from VSAC.
Changes to handle the scenario when the 8 hour ticket from VSAC seems to
expire before the expiration time while trying to retrieve the service
ticket and the response comes up as 401 unauthorized.
Changes to remove the 8 hour ticket from session and provide message to
the user to log into UMLS again.
Added the checks and the error messages in the following screens:
1. Upload to Bonnie screen
2. Retrieval of a Value Set and Update of Value Sets in both the Measure
and Library screens
3. Retrieval of a direct reference code in both the Measure and Library
screens
4. Measure Packager screen when trying to Package and Upload to Bonnie.

Former-commit-id: 2aee9ac070d4b9a35114cbbf1a75656e6f11d9b8