Renamed variables related to HR generation
Measure Bundle is an overloaded term in the MAT.
Often, it refers to the Bundle of artifacts generated by the MAT's
export function and saved off as a zip file. In FHIR, it is a
container of FHIR Resources.

Human Readable generation uses the FHIR Bundle Resource to
generate an artifact that will included in the MAT's Measure
Bundle.

Renamed the List of data requirements filtered to ones
with no backing value set or code to better reflect the check is
against whether the data requirement has a populated code filter,
which contains either the value set, code, or neither.