Style the checkboxes
Indent the match lib name checkbox to show its dependency on the
generate cms ID checkbox.

Indention is done by placing an unordered list with the child checkbox in a HTMLPanel.
Other options, such as UiBinder or building a custom widget,
seemed overkill for a simple indent.
Alas, GWT does not have a unordered/ordered list widget.

Collapsed the margins between the checkboxes and adjusted the
padding in the unordered list to help them look like they're related.