Ignore field level validations when textbox is disabled
When the option to generate the lib name is selected,
the CQL Library Name textbox will be disabled and the field level
check can be safely skipped.

The name will be generated, and validated, by the system on save.
It is reasonable to assume the name will be valid. There is no error
message associated with invalid generated library names.