Remove repeating sys out statements
GWT can be tricky to log from shared, meaning used in both the transpiled JS and the backend services, classes. Sys out prints can ease this issue, but have resulted in overlogging in some places.

Removing sys outs that are not commonly considered when reviewing logs and , thus, are causing fluff to read past.