Update the HQMF QDM Data Type templates to match the CQL-based HQMF STU4.1 IG
The HQMF export serves as an XML based representation of the eCQM andq is based on the CQL Based HQMF specification.
Each update to QDM will usually be followed by an update to the HQMF spec adding support for the new QDM version.

Here the HQMF export is being updated with changes made in STU4.1, which are mostly updated version value for several QDM Data Type Templates,
which are the HQMF's way of identifying a Data Type and adding/removing QDM data types per QDM 5.6 changes.

See the HL7 IG for CQL-based HQMF STU 4.1 Vol 3 chapter 9 for Templates changes.