Rephrase Title and Screenshot requests in PR template
The PR template's ask that submitters provide jira and desc in the Title field was confusing. Devs were adding the jira number to the top of the PR body instead of the title.

Rephrasing to emphasize the jira number should go in the title field, and removed whitespace between the top and the body description section.

Comment clarifying the Screenshot section has a `N/A` box that should be checked if Screenshots are not being included, and to delete the `N/A` line if screenshots are included.