Hack to handle circular dependency between HAPI and the Human Readable
When generating a Human Readable for FHIR measures, MAT relies on the terminologies stored in HAPI. When storing a measure in HAPI, to be cqfm-publishable compliant, the measure needs its text element populated, which uses the Human Readable.

As a quick fix, run the push to HAPI twice, once before generating the HR and again right afterwards. This results in an MAT Package containg congruent Human Readable and Measure Resource with its text element populated.