Correctly permit the deletion of FHIR DRCs associated with versionless code systems
Updating DRC handling to check for null code system version before attempting to extract the value from a non-existent element in the measure xml.

This is currently a FHIR only issue, as QDM measures lock DRC changes to the helper UI where the version is maintained in the backing model.

For both FHIR and QDM, when a code system is added to the CQL via the helper UI (by adding a code), its version is maintained in the model regardless of the "include version" option even though the generated CQL will not contain the version if the "include version" option was not selected. For FHIR, saving from the CQL Editor overwrites the DRCs and Code Systems in the back model. Therefore, the backing model will lose the code system version value whenever CQL without the version component is saved from the CQL Editor. Pre-existing code handling always assumed the version value would be present in the backing model, causing NPEs on delete and add when a DRC with a versionless CS was already present.

Signed-off-by: Joseph Kotanchik <joseph.kotanchik@semanticbits.com>