Add referenced Defines and Functions from external libraries to FHIR Human Readable
When a define or function used in the call stack for a population (IPP, Denom, Num, etc) is from an external library (Global, FHIRHelpers, etc), include that define or function in the FHIR Human Readable.

Support for external defines/functions already exists for QDM HRs. This change brings the definition and function lists in the FHIR HR up to parity with QDM HR.

The current FHIR specific logic for generating the list of defines/functions used by the populations does not account for external libraries. Since the similar QDM logic does account for external libraries and is actually model agnostic, remove the FHIR specific logic and use the existing QDM logic in its place.

The call to the FHIR Microservices is still necessary to ensure validation is complete prior to assembling the Human Readable, however; non-exception results can now be safely ignored.

Signed-off-by: Joseph Kotanchik <joseph.kotanchik@semanticbits.com>