Build the logout redirect URL from Href
The GWT `createUrlBuilder()` method functions more as a find and replace against the current URL than a new URL builder. This causes any parameters to remain in the built URL, causing issues when logging out from Okta  (where logout URLs must be whitelisted).

Using the Href as a base and stripping away the parameters seems to be the cleanest path to generating a consistent, though environment independent, URL without any parameters or unexpected pathing.

Signed-off-by: Joseph Kotanchik <joseph.kotanchik@semanticbits.com>