Add Tomcat 10 instructions to readme
Tomcat 10 and Spring versions < 6 are non-compatible. To run MAT on Tomcat 10, the generated WAR must be run through the Tomcat Migration Tool for Jakarta EE.

This is easily accomplished on Tomcat 10 directly by copying the WAR into the `webapps-javaee` directory. From there, Tomcat will pick up the WAR, run it against the migration tool, and move the result to the normal `webapps` dir.

I have yet to find a way to instruct IntelliJ to deploy to specific directory. As a workaround, add the migration tool execution as a pre-deployment step to the Run config.

Adding steps to the README.md to guide developers in setting up the pre-deployment step that runs migration tool and replaces the target WAR with its result.

Signed-off-by: Joseph Kotanchik <joseph.kotanchik@semanticbits.com>