Use the file naming format from export files in the HQMF's relatedDocuments section
For the URL used, the root is statically set to the MAT base url and the first part of the path is the measure's setId.

This updates the second part of the URL's path to follow the same naming conventions as the export file names (`libraryName`-v`libVersion`-`model`-`modelVersion`.`fileExt`) with dashes swapped in place of all periods and underscores.

Signed-off-by: Joseph Kotanchik <joseph.kotanchik@semanticbits.com>