Set the Organization OID as the Steward ID in the SimpleXML
Goal: Use Organization OID in HQMF's custodian field.

As of commit date, ~2.75 years ago the SimpleXML generator was updated
to set the Steward ID with our unique generated ID for the organization.

Unexpectedly, since the SimpleXML is used for so many things, it also
impacted the generated HQMF by supplying the organization's unique ID
for the custodian field when the OID was expected.